check_entry_proxy: move functions to checker.go
